Fix RMS service API, several dfu issues and some compiler warnings

+++ cliutl/src/setwork.c
@@ -22,8 +22,7 @@ int SYS$EXPREG(int pagcnt, long ret[2], int a, int b);
 set_working_set(int argc, char**argv)
 {
     int pid = 1;
-
-    unsigned long sts;
+    int sts;
 
     $DESCRIPTOR(p, "p1");
     $DESCRIPTOR(d, "quota");
